Replaced "compile" with a simpler "decorate" command
The new command will no longer copy files around and it will simply
generate the manifest and .debug files in the build "source" folder.

Rationale: the "compile" command was a leftover of the "csxs" command
line tool upon which "grunt-cep" was originally built. While in csxs the
command effectively compiled something (i.e. ActionScript files into an
SWF file) amongst other things, in grunt-cep it only performed two
different, mostly unrelated actions: it copied all source files into a
folder of choice and it created manifest files.
Since chances are very high that the source files feeded to cepy are the
result of a previous build process, manifest files are now always
generated in the build "source" folder. If a project needs to perform a
copy of the source files before executing cepy, there are better tools
to do that.