Introduced support to bundle additional files
Similar to what it's currently possible in grunt-cep. If the additional
files require some special setup (such as e.g. Photoshop scripts), a
custom MXI template needs to be provided with the appropriate <File>
tags in place.

An attempt was made to generate those tags automatically, but the
possibilities are simply too many - and the config file will get so
complex that is easier to manually specify the tags in a custom MXI
template.